Avoid partials map for non-partial objects
Fixnums, true, false, and nil cannot be partially unmarshaled and
are never registered with the partial map, but the previous code
still attempted to remove them from that map. Likely due to
overhead from acquiring and installing an identity hash code in
RubyFixnum objects, this appears to have significantly degraded
Marshal.load performance for large numbers of int-ranged Integer
objects as reported in jruby/jruby#9136.

The change here removes the "leave" logic from the marshal load
logic for core types which cannot be partial and which do not get
registered with the partial map via an "entry" call. This improves
performance of the reported benchmark back to JRuby 1.7.27 levels.