Avoid partials usage unless there's a proc
The partials map is only used to avoid invoking the caller's proc
with a partially-initialized object, such as one that is still
loading instance variables. If no proc was given, partially-
initialized objects will never be visible, so there's no reason to
track them. This commit eliminates the use of the partials map
completely in that case.

This alone would be sufficient to fix jruby/jruby#9136, but
previous commits also removed the partials usage for core types
that do not get registered in it.

The change here dramatically improves the performance of loading
nearly all types of values in bench/bench_marshal_load.rb and also
reduces allocation by eliminating the partials map when unused.